Re-enable comments route (#31)
* Add authentication

- Build AuthController
- Edit ErrorController but moved error handling back into each Controller as quick bug fix
- Improve nominationController /all/recipient/:id to better handle errors
- Add /one/email/ route to UserController for authentication
- Add verbose mongoose connection logging
- Create test for password hashing of seed users
- Allowed null for value of release date in NominationModel
- prevent passwordHash being selected automatically in UserModel
- Correct DropDBAll
- Add password hashing to SeedAll for seed data in db
- Start SeedOneUser as a quick method to add a user in the backend
- improved readability to server.js
- Add jasonwebtoken

* bugfix: backtracking

* Update AuthController.js

* bug fix - API crashing

- fixed UserController route /one/email/ by adding :email

* Token auth on all relevant routes

- Create verifyToken to check JWT token
- Correct AuthController, and tidy up code
- NominationController: introduce auth to all relevant routes
- UserController: introduce auth to all relevant routes
- NominationModel - tidy up comments and add requirements on fields
- UserModel - tidy up comments and add requirements on fields

* Rerturn user details on login

- Add return of user details to token when logging in

* Update AuthController.js

add id to response json

* auth ^2

- checked authentication on routes
- added authorisation on relevant routes
- added authorisation test to /users/all to only send a message, as a test
- required further testing.

* Comments Routes

- Create CommentModel
- Create partial CommentController
- Recreate ErrorController - later move error handling to middleware for DRY code
- Create SeedComment

* Update CommentController.js

Big fix - missing close bracket.

* Password Reset

- add PATCH /auth/reset:id route for users to change their own passwords, and admins to change passwords
- amend auth/login
- removed commenting from CommentController
- corrected undefined results in errorController
- remove comments from NominationController
- remove comments from UserController
- remove comments from database.js

* bugfix: UserController

- correct update/admin and /update/self PATCH logic
- Add 401 to error handling

* Update AuthController.js

* User Validation

- import Joi
- Create UserValidation.js
- Add validation to POST /users/new
- Add validation to PATCH /users/self/update/:id
-Add validation to PATCH /users/admin/update/:id
- Move procfile
- Remove image handling
- Drink more coffee

* Nomination Validation

- add validation to NominationController
- add Joi date
- move validation to within src

* bugfix: validation error on releaseDate

- fix validation on releaseDate for Nomination by removing superfluous validation from NominationModel
- install @types/jest and @types/supertest
- Remove commented code from NominationController
- Remove commented code from UserController
- Remove LogToFile from database.js
- Remove LogToFile from index.js
- Rework error message response on server.js
 - Purge incorrect copy of UserValidation

* CommentController and Seed extension

- Remove flat and lodash
- Rework CommentController
- Reorder GET /all/nominator/:firstName/:lastName route in NominationController.js
- Correct HTTP error status codes
- Add users to seed.js
- Temporarily comment out /comments path in server.js

* Nomination pagenation

- Add nomination pagenation

* Bugfix and Code tidy up

- Change Joi validation for updateNominationSchema of nominationID from .required() to .optional()
- add optional to releaseDate in same function for explicitness
- Remove logToFile from seed.js and ErrorController.js
- Remove UPN from seed.js, UserModel.js
- Improve consistency is error handling across UserController, CommentController and NominationController
- Delete unused ImageController

* bugfix

- Security enhancement on database.js
- Add new comment validation

* enable Comments routes

- re-enabled comments/ routes
- add "//" comments for comments routes
- Edit error message in SeedComment.js

+++ src/controllers/CommentController.js
@@ -27,6 +27,7 @@ router.get('/all', auth, async (request, response) => {
 });
 
 // GET all comments on a nomination
+// eg GET localhost:3000/comments/all/nomination/5f2f8e3d2b8e9a0017b0e9f0
 router.get('/all/nomination/:id', async (request, response) => {
   try {
     const result = await Comment.find({ nominationId: request.params.id });
@@ -37,6 +38,7 @@ router.get('/all/nomination/:id', async (request, response) => {
 });
 
 // GET one comment by id
+// eg GET localhost:3000/comments/one/nomination/5f2f8e3d2b8e9a0017b0e9f0
 router.get('/one/nomination/:id', async (request, response) => {
   try {
     const result = await Comment.findById(request.params.id);
@@ -47,6 +49,7 @@ router.get('/one/nomination/:id', async (request, response) => {
 });
 
 // GET all comments by poster user id
+// eg GET localhost:3000/comments/all/user/5f2f8e3d2b8e9a0017b0e9f0
 router.get('/all/user/:id', async (request, response) => {
   const _id = request.params.id;
   try {
@@ -62,6 +65,7 @@ router.get('/all/user/:id', async (request, response) => {
 
 
 // POST a comment on a nomination
+// eg POST localhost:3000/comments/post/5f2f8e3d2b8e9a0017b0e9f0
 router.post('/post/:id',auth, async (request, response) => {
   const _id = request.userId;
 
@@ -86,6 +90,7 @@ router.post('/post/:id',auth, async (request, response) => {
 
 // POST Return all comments on an array of nominations
 // Requires a request with array of nomination ids
+// eg POST localhost:3000/comments/all/nominations
 router.post('/all/nominations', async (request, response) => {
   try {
     const nominationIds = request.body.nominationIds;
@@ -101,6 +106,7 @@ router.post('/all/nominations', async (request, response) => {
 
 
 // PATCH a comment by comment id
+// eg PATCH localhost:3000/comments/update/5f2f8e3d2b8e9a0017b0e9f0
 router.patch('/update/:id', async (request, response) => { 
   // verify user made the comment
   const _id = request.userId;
@@ -122,14 +128,15 @@ router.patch('/update/:id', async (request, response) => {
 
 
 // DELETE a comment by id
+// eg DELETE localhost:3000/comments/delete/5f2f8e3d2b8e9a0017b0e9f0
 router.delete('/delete/:id', async (request, response) => {
   // const _id = request.userId;
 
   try {
-    // const result = await User.findById(_id);
-    // if (!result.isAdmin) {
-    //   return response.status(403).send('You are not authorised to do that. Wash your mouth with soap.');
-    // }
+    const result = await User.findById(_id);
+    if (!result.isAdmin) {
+      return response.status(403).send('You are not authorised to do that. Wash your mouth with soap.');
+    }
 
     const outcome = await Comment.findByIdAndDelete(request.params.id);
